move all csv config to its own trait
