+++ src/CsvControlTrait.php
@@ -1,0 +1,627 @@
+<?php
+/**
+* This file is part of the League.csv library
+*
+* @license http://opensource.org/licenses/MIT
+* @link https://github.com/thephpleague/csv/
+* @version 9.0.0
+* @package League.csv
+*
+* For the full copyright and license information, please view the LICENSE
+* file that was distributed with this source code.
+*/
+declare(strict_types=1);
+
+namespace League\Csv;
+
+use CallbackFilterIterator;
+use InvalidArgumentException;
+use LimitIterator;
+use LogicException;
+use OutOfBoundsException;
+use SplFileObject;
+
+/**
+ *  An abstract class to enable basic CSV manipulation
+ *
+ * @package League.csv
+ * @since  9.0.0
+ *
+ */
+trait CsvControlTrait
+{
+    /**
+     * The path
+     *
+     * can be a StreamIterator object, a SplFileObject object or the string path to a file
+     *
+     * @var StreamIterator|SplFileObject|string
+     */
+    protected $path;
+
+    /**
+     * The file open mode flag
+     *
+     * @var string
+     */
+    protected $open_mode;
+
+    /**
+     * the field delimiter (one character only)
+     *
+     * @var string
+     */
+    protected $delimiter = ',';
+
+    /**
+     * the field enclosure character (one character only)
+     *
+     * @var string
+     */
+    protected $enclosure = '"';
+
+    /**
+     * the field escape character (one character only)
+     *
+     * @var string
+     */
+    protected $escape = '\\';
+
+    /**
+     * newline character
+     *
+     * @var string
+     */
+    protected $newline = "\n";
+
+    /**
+     * Charset Encoding for the CSV
+     *
+     * @var string
+     */
+    protected $input_encoding = 'UTF-8';
+
+    /**
+     * The Input file BOM character
+     * @var string
+     */
+    protected $input_bom;
+
+    /**
+     * The Output file BOM character
+     * @var string
+     */
+    protected $output_bom = '';
+
+    /**
+     * collection of stream filters
+     *
+     * @var array
+     */
+    protected $stream_filters = [];
+
+    /**
+     * Stream filtering mode to apply on all filters
+     *
+     * @var int
+     */
+    protected $stream_filter_mode = STREAM_FILTER_ALL;
+
+    /**
+     *the real path
+     *
+     * @var string the real path to the file
+     *
+     */
+    protected $stream_uri;
+
+    /**
+     * PHP Stream Filter Regex
+     *
+     * @var string
+     */
+    protected $stream_regex = ',^
+        php://filter/
+        (?P<mode>:?read=|write=)?  # The resource open mode
+        (?P<filters>.*?)           # The resource registered filters
+        /resource=(?P<resource>.*) # The resource path
+        $,ix';
+
+    /**
+     * validate a string
+     *
+     * @param mixed $str the value to evaluate as a string
+     *
+     * @throws InvalidArgumentException if the submitted data can not be converted to string
+     *
+     * @return string
+     */
+    protected static function validateString($str): string
+    {
+        if (is_string($str) || (is_object($str) && method_exists($str, '__toString'))) {
+            return (string) $str;
+        }
+        throw new InvalidArgumentException('Expected data must be a string or stringable');
+    }
+
+    /**
+     * Sets the field delimiter
+     *
+     * @param string $delimiter
+     *
+     * @throws InvalidArgumentException If $delimiter is not a single character
+     *
+     * @return $this
+     */
+    public function setDelimiter(string $delimiter): self
+    {
+        $this->delimiter = $this->filterControl($delimiter, 'delimiter');
+
+        return $this;
+    }
+
+    /**
+     * Filter Csv control character
+     *
+     * @param string $char Csv control character
+     * @param string $type Csv control character type
+     *
+     * @throws InvalidArgumentException If the Csv control character is not one character only.
+     *
+     * @return string
+     */
+    protected function filterControl(string $char, string $type)
+    {
+        if (1 == strlen($char)) {
+            return $char;
+        }
+
+        throw new InvalidArgumentException(sprintf('The %s must be a single character', $type));
+    }
+
+    /**
+     * Returns the current field delimiter
+     *
+     * @return string
+     */
+    public function getDelimiter(): string
+    {
+        return $this->delimiter;
+    }
+
+    /**
+     * Detect Delimiters occurences in the CSV
+     *
+     * Returns a associative array where each key represents
+     * a valid delimiter and each value the number of occurences
+     *
+     * @param string[] $delimiters the delimiters to consider
+     * @param int      $nb_rows    Detection is made using $nb_rows of the CSV
+     *
+     * @return array
+     */
+    public function fetchDelimitersOccurrence(array $delimiters, int $nb_rows = 1): array
+    {
+        $nb_rows = $this->validateInteger($nb_rows, 1, 'The number of rows to consider must be a valid positive integer');
+        $filter_row = function ($row) {
+            return is_array($row) && count($row) > 1;
+        };
+        $delimiters = array_unique(array_filter($delimiters, function ($value) {
+            return 1 == strlen($value);
+        }));
+        $csv = $this->getIterator();
+        $res = [];
+        foreach ($delimiters as $delim) {
+            $csv->setCsvControl($delim, $this->enclosure, $this->escape);
+            $iterator = new CallbackFilterIterator(new LimitIterator($csv, 0, $nb_rows), $filter_row);
+            $res[$delim] = count(iterator_to_array($iterator, false), COUNT_RECURSIVE);
+        }
+        arsort($res, SORT_NUMERIC);
+
+        return $res;
+    }
+
+    /**
+     * Validate an integer
+     *
+     * @param int    $int
+     * @param int    $minValue
+     * @param string $errorMessage
+     *
+     * @throws InvalidArgumentException If the value is invalid
+     *
+     * @return int
+     */
+    protected function validateInteger(int $int, int $minValue, string $errorMessage): int
+    {
+        if ($int < $minValue) {
+            throw new InvalidArgumentException($errorMessage);
+        }
+
+        return $int;
+    }
+
+    /**
+     * Sets the field enclosure
+     *
+     * @param string $enclosure
+     *
+     * @throws InvalidArgumentException If $enclosure is not a single character
+     *
+     * @return $this
+     */
+    public function setEnclosure(string $enclosure): self
+    {
+        $this->enclosure = $this->filterControl($enclosure, 'enclosure');
+
+        return $this;
+    }
+
+    /**
+     * Returns the current field enclosure
+     *
+     * @return string
+     */
+    public function getEnclosure(): string
+    {
+        return $this->enclosure;
+    }
+
+    /**
+     * Sets the field escape character
+     *
+     * @param string $escape
+     *
+     * @throws InvalidArgumentException If $escape is not a single character
+     *
+     * @return $this
+     */
+    public function setEscape(string $escape): self
+    {
+        $this->escape = $this->filterControl($escape, 'escape');
+
+        return $this;
+    }
+
+    /**
+     * Returns the current field escape character
+     *
+     * @return string
+     */
+    public function getEscape(): string
+    {
+        return $this->escape;
+    }
+
+    /**
+     * Sets the newline sequence characters
+     *
+     * @param string $newline
+     *
+     * @return static
+     */
+    public function setNewline(string $newline): self
+    {
+        $this->newline = (string) $newline;
+
+        return $this;
+    }
+
+    /**
+     * Returns the current newline sequence characters
+     *
+     * @return string
+     */
+    public function getNewline(): string
+    {
+        return $this->newline;
+    }
+    /**
+     * Sets the CSV encoding charset
+     *
+     * @param string $str
+     *
+     * @return static
+     */
+    public function setInputEncoding(string $str): self
+    {
+        $str = str_replace('_', '-', $str);
+        $str = filter_var($str, FILTER_SANITIZE_STRING, ['flags' => FILTER_FLAG_STRIP_LOW | FILTER_FLAG_STRIP_HIGH]);
+        if (empty($str)) {
+            throw new InvalidArgumentException('you should use a valid charset');
+        }
+        $this->input_encoding = strtoupper($str);
+
+        return $this;
+    }
+
+    /**
+     * Gets the source CSV encoding charset
+     *
+     * @return string
+     */
+    public function getInputEncoding(): string
+    {
+        return $this->input_encoding;
+    }
+
+    /**
+     * Sets the BOM sequence to prepend the CSV on output
+     *
+     * @param string $str The BOM sequence
+     *
+     * @return static
+     */
+    public function setOutputBOM(string $str): self
+    {
+        if (empty($str)) {
+            $this->output_bom = '';
+
+            return $this;
+        }
+
+        $this->output_bom = (string) $str;
+
+        return $this;
+    }
+
+    /**
+     * Returns the BOM sequence in use on Output methods
+     *
+     * @return string
+     */
+    public function getOutputBOM(): string
+    {
+        return $this->output_bom;
+    }
+
+    /**
+     * Returns the BOM sequence of the given CSV
+     *
+     * @return string
+     */
+    public function getInputBOM(): string
+    {
+        if (null === $this->input_bom) {
+            $bom = [
+                AbstractCsv::BOM_UTF32_BE, AbstractCsv::BOM_UTF32_LE,
+                AbstractCsv::BOM_UTF16_BE, AbstractCsv::BOM_UTF16_LE, AbstractCsv::BOM_UTF8,
+            ];
+            $csv = $this->getIterator();
+            $csv->setFlags(SplFileObject::READ_CSV);
+            $csv->rewind();
+            $line = $csv->fgets();
+            $res  = array_filter($bom, function ($sequence) use ($line) {
+                return strpos($line, $sequence) === 0;
+            });
+
+            $this->input_bom = (string) array_shift($res);
+        }
+
+        return $this->input_bom;
+    }
+
+    /**
+     * Internal path setter
+     */
+    protected function initStreamFilter()
+    {
+        if (!is_string($this->path)) {
+            return;
+        }
+
+        if (!preg_match($this->stream_regex, $this->path, $matches)) {
+            $this->stream_uri = $this->path;
+
+            return;
+        }
+
+        $this->stream_uri = $matches['resource'];
+        $this->stream_filters = array_map('urldecode', explode('|', $matches['filters']));
+        $this->stream_filter_mode = $this->fetchStreamModeAsInt($matches['mode']);
+    }
+
+    /**
+     * Get the stream mode
+     *
+     * @param string $mode
+     *
+     * @return int
+     */
+    protected function fetchStreamModeAsInt(string $mode): int
+    {
+        $mode = strtolower($mode);
+        $mode = rtrim($mode, '=');
+        if ('write' == $mode) {
+            return STREAM_FILTER_WRITE;
+        }
+
+        if ('read' == $mode) {
+            return STREAM_FILTER_READ;
+        }
+
+        return STREAM_FILTER_ALL;
+    }
+
+    /**
+     * Check if the trait methods can be used
+     *
+     * @throws LogicException If the API can not be use
+     */
+    protected function assertStreamable()
+    {
+        if (!is_string($this->stream_uri)) {
+            throw new LogicException('The stream filter API can not be used');
+        }
+    }
+
+    /**
+     * Tells whether the stream filter capabilities can be used
+     *
+     * @return bool
+     */
+    public function isActiveStreamFilter(): bool
+    {
+        return is_string($this->stream_uri);
+    }
+
+    /**
+     * stream filter mode Setter
+     *
+     * Set the new Stream Filter mode and remove all
+     * previously attached stream filters
+     *
+     * @param int $mode
+     *
+     * @throws OutOfBoundsException If the mode is invalid
+     *
+     * @return $this
+     */
+    public function setStreamFilterMode(int $mode): self
+    {
+        $this->assertStreamable();
+        if (!in_array($mode, [STREAM_FILTER_ALL, STREAM_FILTER_READ, STREAM_FILTER_WRITE])) {
+            throw new OutOfBoundsException('the $mode should be a valid `STREAM_FILTER_*` constant');
+        }
+
+        $this->stream_filter_mode = $mode;
+        $this->stream_filters = [];
+
+        return $this;
+    }
+
+    /**
+     * stream filter mode getter
+     *
+     * @return int
+     */
+    public function getStreamFilterMode(): int
+    {
+        $this->assertStreamable();
+
+        return $this->stream_filter_mode;
+    }
+
+    /**
+     * append a stream filter
+     *
+     * @param string $filter_name a string or an object that implements the '__toString' method
+     *
+     * @return $this
+     */
+    public function appendStreamFilter(string $filter_name): self
+    {
+        $this->assertStreamable();
+        $this->stream_filters[] = $this->sanitizeStreamFilter($filter_name);
+
+        return $this;
+    }
+
+    /**
+     * prepend a stream filter
+     *
+     * @param string $filter_name a string or an object that implements the '__toString' method
+     *
+     * @return $this
+     */
+    public function prependStreamFilter(string $filter_name): self
+    {
+        $this->assertStreamable();
+        array_unshift($this->stream_filters, $this->sanitizeStreamFilter($filter_name));
+
+        return $this;
+    }
+
+    /**
+     * Sanitize the stream filter name
+     *
+     * @param string $filter_name the stream filter name
+     *
+     * @return string
+     */
+    protected function sanitizeStreamFilter(string $filter_name): string
+    {
+        return urldecode($this->validateString($filter_name));
+    }
+
+    /**
+     * Detect if the stream filter is already present
+     *
+     * @param string $filter_name
+     *
+     * @return bool
+     */
+    public function hasStreamFilter(string $filter_name): bool
+    {
+        $this->assertStreamable();
+
+        return false !== array_search(urldecode($filter_name), $this->stream_filters, true);
+    }
+
+    /**
+     * Remove a filter from the collection
+     *
+     * @param string $filter_name
+     *
+     * @return $this
+     */
+    public function removeStreamFilter(string $filter_name): self
+    {
+        $this->assertStreamable();
+        $res = array_search(urldecode($filter_name), $this->stream_filters, true);
+        if (false !== $res) {
+            unset($this->stream_filters[$res]);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Remove all registered stream filter
+     *
+     * @return $this
+     */
+    public function clearStreamFilter(): self
+    {
+        $this->assertStreamable();
+        $this->stream_filters = [];
+
+        return $this;
+    }
+
+    /**
+     * Return the filter path
+     *
+     * @return string
+     */
+    protected function getStreamFilterPath(): string
+    {
+        $this->assertStreamable();
+        if (!$this->stream_filters) {
+            return $this->stream_uri;
+        }
+
+        return 'php://filter/'
+            .$this->getStreamFilterPrefix()
+            .implode('|', array_map('urlencode', $this->stream_filters))
+            .'/resource='.$this->stream_uri;
+    }
+
+    /**
+     * Return PHP stream filter prefix
+     *
+     * @return string
+     */
+    protected function getStreamFilterPrefix(): string
+    {
+        if (STREAM_FILTER_READ == $this->stream_filter_mode) {
+            return 'read=';
+        }
+
+        if (STREAM_FILTER_WRITE == $this->stream_filter_mode) {
+            return 'write=';
+        }
+
+        return '';
+    }
+}
